feat: add redocBundleURL option
It is helpful to be able render documentation with a non-standard
redoc bundle, e.g., for testing.  In theory, one could use a custom
HTML template for this purpose but in practice that's not
straight-forward because the template would have to be different
depending on whether or not `preview-docs` or `build-docs` is run.
The template would also have to include all the CSS code.